Remove "Backport" from :rx-backport file names
