+++ rx-backport/src/test/java/drewhamilton/skylight/backport/rx/RxSkylightForCoordinatesTest.kt
@@ -3,27 +3,25 @@ package drewhamilton.skylight.backport.rx
 import com.nhaarman.mockitokotlin2.any
 import com.nhaarman.mockitokotlin2.doAnswer
 import com.nhaarman.mockitokotlin2.mock
-import drewhamilton.skylight.backport.Coordinates
-import drewhamilton.skylight.backport.SkylightBackport
 import drewhamilton.skylight.backport.SkylightDay
+import drewhamilton.skylight.backport.SkylightForCoordinatesBackport
 import org.junit.Test
 import org.threeten.bp.LocalDate
 import org.threeten.bp.OffsetTime
 import org.threeten.bp.ZoneOffset
 
-class RxSkylightBackportTest {
+class RxSkylightForCoordinatesTest {
 
-    private val dummyCoordinates = Coordinates(50.0, 60.0)
     private val testDawn = OffsetTime.of(12, 0, 0, 0, ZoneOffset.UTC)
 
-    private lateinit var mockSkylight: SkylightBackport
+    private lateinit var mockSkylightForCoordinates: SkylightForCoordinatesBackport
 
     //region getSkylightDaySingle
     @Test
     fun `getSkylightDaySingle emits info and completes`() {
         mockSkylight { date, dawn -> dummyTypical(date, dawn) }
 
-        mockSkylight.getSkylightDaySingle(dummyCoordinates, LocalDate.now()).test()
+        mockSkylightForCoordinates.getSkylightDaySingle(LocalDate.now()).test()
             .assertComplete()
             .assertValueCount(1)
             .assertValueAt(0) { it == dummyTypical(LocalDate.now(), testDawn) }
@@ -35,7 +33,7 @@ class RxSkylightBackportTest {
     fun `getUpcomingSkylightDays emits today's and tomorrow's info and completes`() {
         mockSkylight { date, dawn -> dummyTypical(date, dawn) }
 
-        mockSkylight.getUpcomingSkylightDays(dummyCoordinates).test()
+        mockSkylightForCoordinates.getUpcomingSkylightDays().test()
             .assertComplete()
             .assertValueCount(2)
             .assertValueAt(0) { it == dummyTypical(LocalDate.now(), testDawn) }
@@ -44,9 +42,10 @@ class RxSkylightBackportTest {
     //endregion
 
     private fun mockSkylight(returnFunction: (LocalDate, OffsetTime) -> SkylightDay) {
-        mockSkylight = mock {
-            on { getSkylightDay(any(), any()) } doAnswer { invocation ->
-                returnFunction(invocation.getArgument(1), testDawn)
+        mockSkylightForCoordinates = mock {
+            on { getSkylightDay(any<LocalDate>()) } doAnswer { invocation ->
+                val date: LocalDate = invocation.getArgument(0)
+                returnFunction(date, testDawn)
             }
         }
     }
