+++ rx-backport/src/test/java/drewhamilton/skylight/backport/rx/RxSkylightTest.kt
@@ -3,25 +3,27 @@ package drewhamilton.skylight.backport.rx
 import com.nhaarman.mockitokotlin2.any
 import com.nhaarman.mockitokotlin2.doAnswer
 import com.nhaarman.mockitokotlin2.mock
+import drewhamilton.skylight.backport.Coordinates
+import drewhamilton.skylight.backport.SkylightBackport
 import drewhamilton.skylight.backport.SkylightDay
-import drewhamilton.skylight.backport.SkylightForCoordinatesBackport
 import org.junit.Test
 import org.threeten.bp.LocalDate
 import org.threeten.bp.OffsetTime
 import org.threeten.bp.ZoneOffset
 
-class RxSkylightForCoordinatesBackportTest {
+class RxSkylightTest {
 
+    private val dummyCoordinates = Coordinates(50.0, 60.0)
     private val testDawn = OffsetTime.of(12, 0, 0, 0, ZoneOffset.UTC)
 
-    private lateinit var mockSkylightForCoordinates: SkylightForCoordinatesBackport
+    private lateinit var mockSkylight: SkylightBackport
 
     //region getSkylightDaySingle
     @Test
     fun `getSkylightDaySingle emits info and completes`() {
         mockSkylight { date, dawn -> dummyTypical(date, dawn) }
 
-        mockSkylightForCoordinates.getSkylightDaySingle(LocalDate.now()).test()
+        mockSkylight.getSkylightDaySingle(dummyCoordinates, LocalDate.now()).test()
             .assertComplete()
             .assertValueCount(1)
             .assertValueAt(0) { it == dummyTypical(LocalDate.now(), testDawn) }
@@ -33,7 +35,7 @@ class RxSkylightForCoordinatesBackportTest {
     fun `getUpcomingSkylightDays emits today's and tomorrow's info and completes`() {
         mockSkylight { date, dawn -> dummyTypical(date, dawn) }
 
-        mockSkylightForCoordinates.getUpcomingSkylightDays().test()
+        mockSkylight.getUpcomingSkylightDays(dummyCoordinates).test()
             .assertComplete()
             .assertValueCount(2)
             .assertValueAt(0) { it == dummyTypical(LocalDate.now(), testDawn) }
@@ -42,10 +44,9 @@ class RxSkylightForCoordinatesBackportTest {
     //endregion
 
     private fun mockSkylight(returnFunction: (LocalDate, OffsetTime) -> SkylightDay) {
-        mockSkylightForCoordinates = mock {
-            on { getSkylightDay(any<LocalDate>()) } doAnswer { invocation ->
-                val date: LocalDate = invocation.getArgument(0)
-                returnFunction(date, testDawn)
+        mockSkylight = mock {
+            on { getSkylightDay(any(), any()) } doAnswer { invocation ->
+                returnFunction(invocation.getArgument(1), testDawn)
             }
         }
     }
